refactor: UserFormController의 메서드 UserController로 이동
UserController와 UserFormController의 private 메서드 일부가 겹치는 데다가,
user의 정보와 관련된 페이지를 UserController에서 제공해주어도 좋을 것 같아 이동하였습니다.
